requirements.txt
docopt==0.6.2
matplotlib==3.2.1
numpy==1.18.2
opencv-python==4.2.0
requests==2.23.0
scipy==1.4.1
svg.path==4.0.2
urllib3==1.25.9
